requirements.txt
iotconnect-sdk-lite
paho-mqtt
requests
